+++ example/bug-fixed/tests-compile-commands-json/a.cc
@@ -1,0 +1,5 @@
+#include <iostream>
+
+int main() {
+    std::cout << "a.cc run successfully!" << std::endl;
+}