+++ tests/projects/c/cosmocc/static_library/xmake.lua
@@ -3,13 +3,13 @@ add_rules("mode.release", "mode.debug")
 add_requires("cosmocc")
 set_toolchains("@cosmocc")
 
-target("foo")
+target("foo", function()
     set_kind("static")
     add_files("src/foo.c")
+end)
 
-target("demo")
+target("demo", function()
     set_kind("binary")
     add_deps("foo")
     add_files("src/main.c")
-
-
+end)